Move thread_rng and co to new module
Rationale: lib.rs should not contain a big mess of unrelated code
